Adds support to format to AsciiDoc
This enables support for AsciiDoc output.

This will generate multiple output files that are linked.

- release-notes.asciidoc
	- release-notes-VERSION.asciidoc
		- release-notes-VERSION-human.asciidoc
		- release-notes-VERSION-generated.asciidoc

If VERSION is e.g 7.12.4 in the overview all versions down to 7.12.0 are
linked highest to lowest. If no version specific notes exists for the previous versions
they will link to github releases.

The human files are empty by default and only generated if they don't
exists. This allows you to write human readable highlights in addition
to the generated notes.